requirements.txt
Flask==2.0.1
requests==2.31.0
python-dotenv==1.0.0
openai==1.68.2
twilio==8.10.2
pymongo==4.5.0
edge-tts==6.1.9
asyncio==3.4.3
